reorganize plugins to simplify import strategy
also update travis and setup

travis removed hunspell dependency since pyontutils is not being
installed with the spell option
